removing boolean converters, adding type consts, renaming TabList to Tabs.
