--- packages/web-components/src/tabs/tab-list/index.ts
@@ -1,5 +1,0 @@
-export * from './tab-list.js';
-export * from './tab-list.options.js';
-export { template as TabListTemplate } from './tab-list.template.js';
-export { styles as TabListStyles } from './tab-list.styles.js';
-export { definition as TabListDefinition } from './tab-list.definition.js';
