--- packages/web-components/src/tabs/tab-list/tab-list.options.ts
@@ -1,15 +1,0 @@
-export const TabListAppearance = {
-  subtle: 'subtle',
-  transparent: 'transparent',
-} as const;
-
-export const TabListSize = {
-  small: 'small',
-  medium: 'medium',
-  large: 'large',
-} as const;
-
-export const TabListOrientation = {
-  vertical: 'vertical',
-  horizontal: 'horizontal',
-} as const;
