+++ packages/web-components/src/tabs/tabs/define.ts
@@ -1,4 +1,4 @@
 import { FluentDesignSystem } from '../../fluent-design-system.js';
-import { definition } from './tab-list.definition.js';
+import { definition } from './tabs.definition.js';
 
 definition.define(FluentDesignSystem.registry);
