+++ packages/web-components/src/tabs/tabs/index.ts
@@ -1,0 +1,5 @@
+export * from './tabs.js';
+export * from './tabs.options.js';
+export { template as TabListTemplate } from './tabs.template.js';
+export { styles as TabListStyles } from './tabs.styles.js';
+export { definition as TabListDefinition } from './tabs.definition.js';
