+++ packages/web-components/src/tabs/tabs/tabs.spec.md
@@ -1,4 +1,4 @@
-# TabList
+# Tabs
 
 ## Description
