+++ packages/web-components/src/tabs/tabs/tabs.template.ts
